+++ docs/src/changelog.md
@@ -29,9 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Total water storage as an export variable for `SBM` concept. This is the total water stored
   per grid cell in millimeters. Excluded from this variable are the floodplain, lakes and
   reservoirs.
-- Checks to see if all states are covered in the .toml. If not all states are covered, an error
-  is thrown. If there are more states then required specified, these states are ignored (with a
-  warning in the logging), and the simulation will continue.
+- Checks to see if all states are covered in the .toml file. If not all states are covered,
+  an error is thrown. If there are more states specified than required, these states are
+  ignored (with a warning in the logging), and the simulation will continue.
 
 ## v0.7.3 - 2024-01-12